// src/interfaces/Interfaces.tsx
export interface SinglePokemon {
  id: string;
  name: string;
  picture: string;
  color?: string;
}// Generated by https://quicktype.io